refactor: move HoverAdapter to new file
