+++ src/language/common/lspLanguageFeatures.ts
@@ -7,7 +7,6 @@ import * as lsTypes from 'vscode-languageserver-types';
 import {
 	languages,
 	editor,
-	IMarkdownString,
 	Uri,
 	Position,
 	Range,
@@ -23,77 +22,7 @@ export * from './DiagnosticsAdapter';
 
 export * from './CompletionAdapter';
 
-//#region HoverAdapter
-
-export interface ILanguageWorkerWithHover {
-	doHover(uri: string, position: lsTypes.Position): Promise<lsTypes.Hover | null>;
-}
-
-export class HoverAdapter<T extends ILanguageWorkerWithHover> implements languages.HoverProvider {
-	constructor(private readonly _worker: WorkerAccessor<T>) {}
-
-	provideHover(
-		model: editor.IReadOnlyModel,
-		position: Position,
-		token: CancellationToken
-	): Promise<languages.Hover | undefined> {
-		let resource = model.uri;
-
-		return this._worker(resource)
-			.then((worker) => {
-				return worker.doHover(resource.toString(), fromPosition(position));
-			})
-			.then((info) => {
-				if (!info) {
-					return;
-				}
-				return <languages.Hover>{
-					range: toRange(info.range),
-					contents: toMarkedStringArray(info.contents)
-				};
-			});
-	}
-}
-
-function isMarkupContent(thing: any): thing is lsTypes.MarkupContent {
-	return (
-		thing && typeof thing === 'object' && typeof (<lsTypes.MarkupContent>thing).kind === 'string'
-	);
-}
-
-function toMarkdownString(entry: lsTypes.MarkupContent | lsTypes.MarkedString): IMarkdownString {
-	if (typeof entry === 'string') {
-		return {
-			value: entry
-		};
-	}
-	if (isMarkupContent(entry)) {
-		if (entry.kind === 'plaintext') {
-			return {
-				value: entry.value.replace(/[\\`*_{}[\]()#+\-.!]/g, '\\$&')
-			};
-		}
-		return {
-			value: entry.value
-		};
-	}
-
-	return { value: '```' + entry.language + '\n' + entry.value + '\n```\n' };
-}
-
-function toMarkedStringArray(
-	contents: lsTypes.MarkupContent | lsTypes.MarkedString | lsTypes.MarkedString[]
-): IMarkdownString[] | undefined {
-	if (!contents) {
-		return void 0;
-	}
-	if (Array.isArray(contents)) {
-		return contents.map(toMarkdownString);
-	}
-	return [toMarkdownString(contents)];
-}
-
-//#endregion
+export * from './HoverAdapter';
 
 //#region DocumentHighlightAdapter
